Track project-specific style configs
These set of configs should be maximally compatible with professor Pearce's demo code.
That is, reformatting them with IDEA should introduce the least amount of changes.

Unfortunately IDEA is not very flexible so there are still going to be some changes,
for example it's not possible to have an empty method body written as `{}`, while one-line
methods written as `{ return 0; }` (with spaces).